update: nest - use next/navigation
