+++ package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "@next/font": "13.0.7",
+    "@portabletext/react": "^3.2.1",
     "@portabletext/to-html": "^1.0.4",
     "@types/node": "18.11.17",
     "@types/react-dom": "^18.3.1",
@@ -21,7 +22,7 @@
     "eslint": "^9.13.0",
     "eslint-config-next": "^15.0.1",
     "fast-xml-parser": "^4.5.0",
-    "html-react-parser": "^3.0.6",
+    "html-react-parser": "^3.0.16",
     "next": "^15.2.4",
     "next-sanity": "^9.8.7",
     "react": "^18.3.1",
