+++ src/app/globals.css
@@ -1,4 +1,4 @@
-@import url('components.css');
+@import url('../../styles/components.css');
 
 :root {
   /* colors light theme(default theme) */
