+++ src/app/layout.tsx
@@ -1,0 +1,51 @@
+import type { Metadata } from "next";
+import { Geist, Geist_Mono } from "next/font/google";
+
+import "./globals.css";
+import { ThemeProvider } from "@/components/ThemeProvider";
+import Header from "@/components/header";
+import Footer from "@/components/Footer";
+
+const geistSans = Geist({
+    variable: "--font-geist-sans",
+    subsets: ["latin"],
+});
+
+const geistMono = Geist_Mono({
+    variable: "--font-geist-mono",
+    subsets: ["latin"],
+});
+
+export const metadata: Metadata = {
+    title: "Improvised JI..",
+    description: "Improvised Project Management system",
+};
+
+export default function RootLayout({
+    children,
+}: Readonly<{
+    children: React.ReactNode;
+}>) {
+    return (
+        <html lang="en" >
+            <head>
+                <title>Ricardo, Software Engineer</title>
+                <meta
+                    name="description"
+                    content="Ricardo, software engineer personal site"
+                />
+                <meta name="viewport" content="width=device-width, initial-scale=1" />
+            </head>
+            <body>
+                <ThemeProvider>
+                    <Header
+                    />
+                    <main>
+                        {children}
+                    </main>
+                </ThemeProvider>
+                <Footer />
+            </body>
+        </html>
+    );
+}
