+++ src/app/mobile/page.tsx
@@ -1,0 +1,5 @@
+import MobileView from "../../../views/mobile"
+
+export default function Mobile() {
+  return <MobileView />
+}
