+++ utils/api.ts
@@ -1,7 +1,7 @@
 import axios from "axios";
 
 const api = {
-    getResumeData: () => axios.get(`http://localhost:3000/api/resume`)
+    getResumeData: () => axios.get<any>(`http://localhost:3000/api/resume`).then(response => response.data)
 }
 
 export default api;
