+++ utils/contexts.ts
@@ -1,3 +1,4 @@
+"use client";
 import React from "react";
 
-export const ThemeModeContext = React.createContext({ darkModeEnabled: false });
+export const ThemeModeContext = React.createContext({ darkModeEnabled: false, setDarkModeEnabled: (value: boolean) => { } });
